Make the filename shorter to fit more useful info
Original complaint was that the filename is too long and the error message
is less useful because the filename appears several times and takes up all
space. After 6ff428d the situation got better, but not much.

This commit changes the filename to <tmp>. That's very short but is probably
still clear enough.

Closes #8